merging

+++ osl-graph/osl-graph-api/src/main/java/org/openspotlight/graph/BaseIdentifer.java
@@ -32,6 +32,6 @@ public interface BaseIdentifer extends PropertyContainer {
      * 
      * @return the element weight.
      */
-    BigInteger getWeightValue();
+    int getWeightValue();
 
 }
